+++ src/main/java/umc/spring/service/member/MemberQueryServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
+import org.springframework.transaction.annotation.Transactional;
 import umc.spring.domain.Member;
 import umc.spring.domain.Review;
 import umc.spring.domain.enums.MissionStatus;
@@ -17,18 +18,16 @@ import umc.spring.repository.member.MemberRepository;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class MemberQueryServiceImpl implements MemberQueryService{
 
     private final MemberRepository memberRepository;
     private final ReviewRepository reviewRepository;
-    private final MemberMissionRepository memberMissionRepository;
-
 
     @Override
     public MemberDto findMemberInMyPage(Long memberId) {
-        MemberDto memberDto = memberRepository.findMemberInMyPage(memberId);
 
-        System.out.println(memberDto);
+        MemberDto memberDto = memberRepository.findMemberInMyPage(memberId);
 
         return memberDto;
     }
@@ -36,31 +35,11 @@ public class MemberQueryServiceImpl implements MemberQueryService{
     @Override
     public Page<Review> getMyReviewList(Integer page) {
 
-
+        // 유저 하드 코딩
         Member member = memberRepository.findById(1L).get();
         Page<Review> pageReviewList = reviewRepository.findByMember(member, PageRequest.of(page, 10));
 
         return pageReviewList;
     }
-    @Override
-    public Page<MemberMission> getMyMissionList(Integer page) {
-
-        // 유저 하드 코딩
-        Member member = memberRepository.findById(1L).get();
-        Page<MemberMission> pageMemberMissionList = memberMissionRepository.findByMember(member, PageRequest.of(page, 10));
-
-
-        return pageMemberMissionList;
-    }
-
-    @Override
-    public MemberMission updateMemberMissionStatus(Long memberMissionId) {
-
-        MemberMission memberMission = memberMissionRepository.findById(memberMissionId).get();
-        memberMission.completedMission();
-
-        return memberMissionRepository.save(memberMission);
-    }
-
 
 }
