+++ src/main/java/umc/spring/validation/validator/MemberMissionExist.java
@@ -5,15 +5,14 @@ import jakarta.validation.ConstraintValidatorContext;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import umc.spring.apiPayload.code.status.ErrorStatus;
-import umc.spring.repository.memberMission.MemberMissionRepository;
-import umc.spring.service.memberMission.MemberMissionService;
+import umc.spring.service.memberMission.MemberMissionQueryService;
 import umc.spring.validation.annotation.ExistMemberMission;
 
 @Component
 @RequiredArgsConstructor
 public class MemberMissionExist implements ConstraintValidator<ExistMemberMission, Long> {
 
-    private final MemberMissionService memberMissionService;
+    private final MemberMissionQueryService memberMissionService;
     @Override
     public void initialize(ExistMemberMission constraintAnnotation) {
         ConstraintValidator.super.initialize(constraintAnnotation);
